Refactor i18n modules to fix module_max_lines lint violations
Split oversized modules to comply with the 400-line limit:

- catalog.rs → catalog/mod.rs + catalog/tests.rs
- message.rs → message/mod.rs + message/factories.rs + message/tests.rs

Remove lint suppressions from src/i18n/mod.rs as modules now
comply with module_max_lines rule.
